+++ src/app/clients/clients-view/general-tab/general-tab.component.ts
@@ -1,34 +1,61 @@
 /** Angular Imports */
-import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
+import { Component, OnInit } from "@angular/core";
+import { Router, ActivatedRoute } from "@angular/router";
 
 /** Custom Services. */
-import { ClientsService } from 'app/clients/clients.service';
+import { ClientsService } from "app/clients/clients.service";
 
 /**
  * General Tab component.
  */
 @Component({
-  selector: 'mifosx-general-tab',
-  templateUrl: './general-tab.component.html',
-  styleUrls: ['./general-tab.component.scss']
+  selector: "mifosx-general-tab",
+  templateUrl: "./general-tab.component.html",
+  styleUrls: ["./general-tab.component.scss"],
 })
 export class GeneralTabComponent {
-
   /** Open Loan Accounts Columns */
-  openLoansColumns: string[] = ['Account No', 'Loan Account', 'Original Loan', 'Loan Balance', 'Amount Paid', 'Type', 'Actions'];
+  openLoansColumns: string[] = [
+    "Account No",
+    "Loan Account",
+    "Original Loan",
+    "Loan Balance",
+    "Amount Paid",
+    "Type",
+    "Actions",
+  ];
   /** Closed Loan Accounts Columns */
-  closedLoansColumns: string[] = ['Account No', 'Loan Account', 'Original Loan', 'Loan Balance', 'Amount Paid', 'Type', 'Closed Date'];
+  closedLoansColumns: string[] = [
+    "Account No",
+    "Loan Account",
+    "Original Loan",
+    "Loan Balance",
+    "Amount Paid",
+    "Type",
+    "Closed Date",
+  ];
   /** Open Savings Accounts Columns */
-  openSavingsColumns: string[] = ['Account No', 'Saving Account', 'Last Active', 'Balance', 'Actions'];
+  openSavingsColumns: string[] = ["Account No", "Saving Account", "Last Active", "Balance", "Actions"];
   /** Closed Savings Accounts Columns */
-  closedSavingsColumns: string[] = ['Account No', 'Saving Account', 'Closed Date'];
+  closedSavingsColumns: string[] = ["Account No", "Saving Account", "Closed Date"];
   /** Open Shares Accounts Columns */
-  openSharesColumns: string[] = ['Account No', 'Share Account', 'Approved Shares', 'Pending For Approval Shares', 'Actions'];
+  openSharesColumns: string[] = [
+    "Account No",
+    "Share Account",
+    "Approved Shares",
+    "Pending For Approval Shares",
+    "Actions",
+  ];
   /** Closed Shares Accounts Columns */
-  closedSharesColumns: string[] = ['Account No', 'Share Account', 'Approved Shares', 'Pending For Approval Shares', 'Closed Date'];
+  closedSharesColumns: string[] = [
+    "Account No",
+    "Share Account",
+    "Approved Shares",
+    "Pending For Approval Shares",
+    "Closed Date",
+  ];
   /** Upcoming Charges Columns */
-  upcomingChargesColumns: string[] = ['Name', 'Due as of', 'Due', 'Paid', 'Waived', 'Outstanding', 'Actions'];
+  upcomingChargesColumns: string[] = ["Name", "Due as of", "Due", "Paid", "Waived", "Outstanding", "Actions"];
 
   /** Client Account Data */
   clientAccountData: any;
@@ -62,20 +89,16 @@ export class GeneralTabComponent {
    * @param {ClientsService} clientService Clients Service
    * @param {Router} router Router
    */
-  constructor(
-    private route: ActivatedRoute,
-    private clientService: ClientsService,
-    private router: Router
-  ) {
-    this.route.data.subscribe((data: { clientAccountsData: any, clientChargesData: any, clientSummary: any }) => {
+  constructor(private route: ActivatedRoute, private clientService: ClientsService, private router: Router) {
+    this.route.data.subscribe((data: { clientAccountsData: any; clientChargesData: any; clientSummary: any }) => {
       this.clientAccountData = data.clientAccountsData;
       this.savingAccounts = data.clientAccountsData.savingsAccounts;
       this.loanAccounts = data.clientAccountsData.loanAccounts;
-     // this.shareAccounts = data.clientAccountsData.shareAccounts;
+      // this.shareAccounts = data.clientAccountsData.shareAccounts;
       this.upcomingCharges = data.clientChargesData.pageItems;
-      this.clientSummary = data.clientSummary ? data.clientSummary[0] : [];
-      this.clientid = this.route.parent.snapshot.params['clientId'];
-  });
+      this.clientSummary = data.clientSummary ? data.clientSummary[0] : null;
+      this.clientid = this.route.parent.snapshot.params["clientId"];
+    });
   }
 
   /**
@@ -119,7 +142,7 @@ export class GeneralTabComponent {
    * @param clientId Selected Client Id.
    */
   waiveCharge(chargeId: string, clientId: string) {
-    const charge = { clientId: clientId.toString(), resourceType: chargeId};
+    const charge = { clientId: clientId.toString(), resourceType: chargeId };
     this.clientService.waiveClientCharge(charge).subscribe(() => {
       this.getChargeData(clientId);
     });
@@ -147,8 +170,10 @@ export class GeneralTabComponent {
    * @param {any} loanId Loan Id
    */
   routeTransferFund(loanId: any) {
-    const queryParams: any = { loanId: loanId, accountType: 'fromloans' };
-    this.router.navigate(['../', 'loans-accounts', loanId, 'transfer-funds', 'make-account-transfer'], { relativeTo: this.route, queryParams: queryParams });
+    const queryParams: any = { loanId: loanId, accountType: "fromloans" };
+    this.router.navigate(["../", "loans-accounts", loanId, "transfer-funds", "make-account-transfer"], {
+      relativeTo: this.route,
+      queryParams: queryParams,
+    });
   }
-
 }