+++ src/app/shared/shared.module.ts
@@ -1,34 +1,35 @@
 /** Angular Imports */
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule, FormsModule } from '@angular/forms';
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { ReactiveFormsModule, FormsModule } from "@angular/forms";
 
 /** Custom Components */
-import { FormfieldComponent } from './form-dialog/formfield/formfield.component';
-import { FormDialogComponent } from './form-dialog/form-dialog.component';
-import { DeleteDialogComponent } from './delete-dialog/delete-dialog.component';
-import { CancelDialogComponent } from './cancel-dialog/cancel-dialog.component';
-import { FileUploadComponent } from './file-upload/file-upload.component';
-import { FooterComponent } from './footer/footer.component';
-import { LanguageSelectorComponent } from './language-selector/language-selector.component';
-import { ThemePickerComponent } from './theme-picker/theme-picker.component';
-import { ChangePasswordDialogComponent } from './change-password-dialog/change-password-dialog.component';
-import { EnableDialogComponent } from './enable-dialog/enable-dialog.component';
-import { DisableDialogComponent } from './disable-dialog/disable-dialog.component';
-import { ConfirmationDialogComponent } from './confirmation-dialog/confirmation-dialog.component';
-import { ErrorDialogComponent } from './error-dialog/error-dialog.component';
-import { NotificationsTrayComponent } from './notifications-tray/notifications-tray.component';
-import { SearchToolComponent } from './search-tool/search-tool.component';
-import { KeyboardShortcutsDialogComponent } from './keyboard-shortcuts-dialog/keyboard-shortcuts-dialog.component';
-import { ServerSelectorComponent } from './server-selector/server-selector.component';
+import { FormfieldComponent } from "./form-dialog/formfield/formfield.component";
+import { FormDialogComponent } from "./form-dialog/form-dialog.component";
+import { DeleteDialogComponent } from "./delete-dialog/delete-dialog.component";
+import { CancelDialogComponent } from "./cancel-dialog/cancel-dialog.component";
+import { FileUploadComponent } from "./file-upload/file-upload.component";
+import { FooterComponent } from "./footer/footer.component";
+import { LanguageSelectorComponent } from "./language-selector/language-selector.component";
+import { ThemePickerComponent } from "./theme-picker/theme-picker.component";
+import { ChangePasswordDialogComponent } from "./change-password-dialog/change-password-dialog.component";
+import { EnableDialogComponent } from "./enable-dialog/enable-dialog.component";
+import { DisableDialogComponent } from "./disable-dialog/disable-dialog.component";
+import { ConfirmationDialogComponent } from "./confirmation-dialog/confirmation-dialog.component";
+import { ErrorDialogComponent } from "./error-dialog/error-dialog.component";
+import { NotificationsTrayComponent } from "./notifications-tray/notifications-tray.component";
+import { SearchToolComponent } from "./search-tool/search-tool.component";
+import { KeyboardShortcutsDialogComponent } from "./keyboard-shortcuts-dialog/keyboard-shortcuts-dialog.component";
+import { ServerSelectorComponent } from "./server-selector/server-selector.component";
 
 /** Custom Modules */
-import { IconsModule } from './icons.module';
-import { MaterialModule } from './material.module';
-import { TranslateModule } from '@ngx-translate/core';
-import { OfficeTreeViewComponent } from './office-tree-view/office-tree-view.component';
-import { OfficeHierarchyComponent } from './office-hierarchy/office-hierarchy.component';
-import { CountryTreeViewComponent } from './country-tree-view/country-tree-view.component';
+import { IconsModule } from "./icons.module";
+import { MaterialModule } from "./material.module";
+import { TranslateModule } from "@ngx-translate/core";
+import { OfficeTreeViewComponent } from "./office-tree-view/office-tree-view.component";
+import { OfficeHierarchyComponent } from "./office-hierarchy/office-hierarchy.component";
+import { CountryTreeViewComponent } from "./country-tree-view/country-tree-view.component";
+import { NgSelectModule } from "@ng-select/ng-select";
 
 /**
  * Shared Module
@@ -36,13 +37,7 @@ import { CountryTreeViewComponent } from './country-tree-view/country-tree-view.
  * Modules and components that are shared throughout the application should be here.
  */
 @NgModule({
-  imports: [
-    CommonModule,
-    IconsModule,
-    MaterialModule,
-    ReactiveFormsModule,
-    TranslateModule.forRoot(),
-  ],
+  imports: [CommonModule, IconsModule, MaterialModule, ReactiveFormsModule, NgSelectModule, TranslateModule.forRoot()],
   declarations: [
     FormfieldComponent,
     FormDialogComponent,
@@ -63,7 +58,7 @@ import { CountryTreeViewComponent } from './country-tree-view/country-tree-view.
     ServerSelectorComponent,
     OfficeTreeViewComponent,
     OfficeHierarchyComponent,
-    CountryTreeViewComponent
+    CountryTreeViewComponent,
   ],
   exports: [
     FileUploadComponent,
@@ -77,11 +72,12 @@ import { CountryTreeViewComponent } from './country-tree-view/country-tree-view.
     CommonModule,
     IconsModule,
     MaterialModule,
+    NgSelectModule,
     FormsModule,
     ReactiveFormsModule,
     OfficeTreeViewComponent,
     OfficeHierarchyComponent,
-    CountryTreeViewComponent
-  ]
+    CountryTreeViewComponent,
+  ],
 })
-export class SharedModule { }
+export class SharedModule {}