+++ src/assets/env.js
@@ -1,20 +1,21 @@
-(function(window) {
+(function (window) {
   window["env"] = window["env"] || {};
 
   // BackEnd Environment variables
-  window["env"]["fineractApiUrls"] = 'https://dev.mifos.io,https://demo.mifos.io,https://qa.mifos.io,https://staging.mifos.io,https://mobile.mifos.io,https://loans.test.oneacrefund.org,https://localhost:8443';
-  window["env"]["fineractApiUrl"]  = 'https://loans.test.oneacrefund.org';
+  window["env"]["fineractApiUrls"] =
+    "https://dev.mifos.io,https://demo.mifos.io,https://qa.mifos.io,https://staging.mifos.io,https://mobile.mifos.io,https://loans.test.oneacrefund.org,https://localhost:8443";
+  window["env"]["fineractApiUrl"] = "https://loans.test.oneacrefund.org";
 
-  window['env']['authServerUrl']  = 'https://accounts.test.oneacrefund.org';
+  window["env"]["authServerUrl"] = "";
 
-  window["env"]["apiProvider"] = '/fineract-provider/api';
-  window["env"]["apiVersion"]  = '/v1';
+  window["env"]["apiProvider"] = "/fineract-provider/api";
+  window["env"]["apiVersion"] = "/v1";
 
-  window["env"]["fineractPlatformTenantId"]  = '';
+  window["env"]["fineractPlatformTenantId"] = "";
 
   // Language Environment variables
-  window["env"]["defaultLanguage"] = '';
-  window["env"]["supportedLanguages"] = '';
+  window["env"]["defaultLanguage"] = "";
+  window["env"]["supportedLanguages"] = "";
   // Head Office ID
-  window["env"]["headOfficeID"] = '';
+  window["env"]["headOfficeID"] = "";
 })(this);