+++ RELEASE_NOTES.md
@@ -33,6 +33,7 @@ Release Notes
     * [SER-2093] - Add qualification rules on loan products    
     * [SER-2396] - Deploy fineract UI v2 on the integration cluster (Added missing envs)
     * [SER-2440] - Revise the country selection option to check for the Head Office instead of roles
+    * [SER-2446] - Add ng-select as droup down in client and group creation
 
 ## Version 1.0.0 - for use with Fineract Web App